added tsconfig for the frontend code + moved all the frontend public stuff in a public folder
