+++ .dockerignore
@@ -1,1 +1,1 @@
-node_modules/
+app/node_modules/
