--- .env.example
@@ -1,2 +1,0 @@
-GOOGLE_OAUTH_CLIENT_ID=corethegame
-GOOGLE_OAUTH_CLIENT_SECRET=iwillnevergiveyouup
