--- .prettierrc.json
@@ -1,8 +1,0 @@
-{
-	"tabWidth": 4,
-	"useTabs": true,
-	"semi": true,
-	"singleQuote": true,
-	"trailingComma": "es5",
-	"printWidth": 80
-}
