+++ app/public/pages/layouts/basic.ejs
@@ -1,63 +1,80 @@
 <!DOCTYPE html>
 <html lang="en">
+	<head>
+		<title>Transcendence: Home</title>
+		<link rel="stylesheet" href="/static/css/styles.css" />
+		<link rel="stylesheet" href="/static/css/popups.css" />
+		<script src="https://cdn.jsdelivr.net/npm/@tailwindcss/browser@4"></script>
+		<link
+			rel="icon"
+			href="/static/assets/favicons/favicon.ico"
+			type="image/x-icon"
+		/>
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<meta
+			name="description"
+			content="Transcendence is a web application for the game of the same name."
+		/>
+	</head>
 
-<head>
-	<title>Transcendence: Home</title>
-	<link rel="stylesheet" href="/public/css/styles.css">
-	<link rel="stylesheet" href="/public/css/popups.css">
-	<script src="https://cdn.jsdelivr.net/npm/@tailwindcss/browser@4"></script>
-	<link rel="icon" href="/public/assets/favicons/favicon.ico" type="image/x-icon">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<meta name="description" content="Transcendence is a web application for the game of the same name.">
-</head>
-
-<body class="bg-gray-100 text-gray-900">
-	<header class="bg-blue-600 text-white p-4 shadow-md">
-		<div class="container mx-auto flex justify-between items-center">
-			<h1 class="text-2xl font-bold" onclick="loadPartialView('index')">Welcome to Transcendence</h1>
-			<div id="menu">
-				<% if (!isAuthenticated) { %>
-					<%- include('layouts/menu/guest') %>
-				<% } else { %>
-					<%- include('layouts/menu/loggedin') %>
-				<% } %>
+	<body class="bg-gray-100 text-gray-900">
+		<header class="bg-blue-600 text-white p-4 shadow-md">
+			<div class="container mx-auto flex justify-between items-center">
+				<h1
+					class="text-2xl font-bold"
+					onclick="loadPartialView('index')"
+				>
+					Welcome to Transcendence
+				</h1>
+				<div id="menu">
+					<% if (!isAuthenticated) { %> <%-
+					include('layouts/menu/guest') %> <% } else { %> <%-
+					include('layouts/menu/loggedin') %> <% } %>
+				</div>
 			</div>
+		</header>
 
-		</div>
-	</header>
-
-	<main class="container mx-auto p-4">
-		<div id="content" class="bg-white p-6 rounded shadow-md">
-			<%- body %>
-		</div>
-		<div id="popup-wrapper" class="fixed bottom-5 left-5 z-50 flex flex-col gap-3 pointer-events-none">
-			<div id="popup-container" class="flex flex-col-reverse gap-3"></div>
-			<button id="close-all-popups-btn"
-				class="mt-2 self-start pointer-events-auto bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">
-				Close all
-			</button>
-		</div>
-	</main>
-
-	<footer class="bg-gray-800 text-white p-4 mt-4">
-		<div class="container mx-auto text-center">
-			<div>
-				<h2>
-					THE HOLY NUMBER OF DESTINY:
-					<span id="numberDisplay"
-						class="select-none cursor-pointer text-xl font-semibold text-blue-400 hover:text-blue-600 transition-colors">
-						0
-					</span>
-					- (why don't you click on it?)
-				</h2>
+		<main class="container mx-auto p-4">
+			<div id="content" class="bg-white p-6 rounded shadow-md">
+				<%- body %>
 			</div>
-			<p>&copy; 2025; 42 Transcendence by jkauker and fschuber</p>
-		</div>
-	</footer>
+			<div
+				id="popup-wrapper"
+				class="fixed bottom-5 left-5 z-50 flex flex-col gap-3 pointer-events-none"
+			>
+				<div
+					id="popup-container"
+					class="flex flex-col-reverse gap-3"
+				></div>
+				<button
+					id="close-all-popups-btn"
+					class="mt-2 self-start pointer-events-auto bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
+				>
+					Close all
+				</button>
+			</div>
+		</main>
 
-	<script src="/static/scripts/script.js" defer></script>
-	<script src="/static/scripts/events.js" defer></script>
-	<script src="/static/scripts/friends.js" defer></script>
-</body>
+		<footer class="bg-gray-800 text-white p-4 mt-4">
+			<div class="container mx-auto text-center">
+				<div>
+					<h2>
+						THE HOLY NUMBER OF DESTINY:
+						<span
+							id="numberDisplay"
+							class="select-none cursor-pointer text-xl font-semibold text-blue-400 hover:text-blue-600 transition-colors"
+						>
+							0
+						</span>
+						- (why don't you click on it?)
+					</h2>
+				</div>
+				<p>&copy; 2025; 42 Transcendence by jkauker and fschuber</p>
+			</div>
+		</footer>
 
+		<script type="module" src="/static/js/script.js" defer></script>
+		<script type="module" src="/static/js/events.js" defer></script>
+		<script type="module" src="/static/js/friends.js" defer></script>
+	</body>
 </html>
