+++ app/public/scripts/add_friends.ts
