+++ app/public/scripts/edit_profile.ts
