+++ app/public/scripts/friends.ts
