+++ app/public/scripts/game.ts
