+++ app/public/scripts/gameRenderer.ts
