+++ app/public/scripts/game_setup.ts
