+++ app/public/scripts/register.ts
