+++ app/public/scripts/script.ts
