+++ app/src/plugins/static.ts
@@ -4,12 +4,7 @@ import path from 'node:path';
 
 export default fp(async (fastify) => {
 	fastify.register(fastifyStatic, {
-		root: path.join(__dirname, '../static/'),
-		prefix: '/static/',
-	});
-	fastify.register(fastifyStatic, {
 		root: path.join(__dirname, '../../public/'),
-		prefix: '/public/',
-		decorateReply: false
+		prefix: '/static/',
 	});
 });
