+++ app/tsconfig.frontend.json
@@ -1,0 +1,12 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"outDir": "public/js",
+		"module": "ESNext",
+		"moduleResolution": "Bundler",
+		"target": "ES2020",
+		"lib": ["DOM", "DOM.Iterable", "ESNext"]
+	},
+	"include": ["public/scripts/**/*.ts"],
+	"exclude": ["node_modules"]
+}
